introduce WritableResourceModuleRegistry
